+++ src/nifgen/metadata/functions_addon.py
@@ -20,8 +20,8 @@ functions_codegen_method = {
     'ConfigureDigitalLevelScriptTrigger':   { 'codegen_method': 'public',   },
     'ConfigureFreqList':                    { 'codegen_method': 'public',   },
     'ConfigureStandardWaveform':            { 'codegen_method': 'public',   },
-    'CreateWaveformF64':                    { 'codegen_method': 'private',  },  # TODO(marcoskirsch): Call this from public method create_waveform()
-    'CreateWaveformI16':                    { 'codegen_method': 'no',       },  # TODO(marcoskirsch): Call this from public method create_waveform()
+    'CreateWaveformF64':                    { 'codegen_method': 'private',  },  # Called from public method create_waveform()
+    'CreateWaveformI16':                    { 'codegen_method': 'private',  },  # Called from public method create_waveform()
     'WriteBinary16Waveform':                { 'codegen_method': 'private',  },  # Called from public method write_waveform()
     'WriteNamedWaveformF64':                { 'codegen_method': 'private',  },  # Called from public method write_waveform()
     'WriteNamedWaveformI16':                { 'codegen_method': 'private',  },  # Called from public method write_waveform()
@@ -130,6 +130,54 @@ functions_default_value = {
 
 # Functions not in original metadata.
 functions_additional_functions = {
+    'CreateWaveformDispatcher': {
+        'codegen_method': 'public',
+        'returns': 'ViStatus',
+        'parameters': [
+            {
+                'direction': 'in',
+                'enum': None,
+                'name': 'vi',
+                'type': 'ViSession',
+            },
+            {
+                'direction': 'in',
+                'enum': None,
+                'name': 'channelName',
+                'type': 'ViConstString',
+            },
+            {
+                'direction': 'in',
+                'enum': None,
+                'name': 'waveformDataArray',
+                'type': 'ViReal64[]',  #TODO(marcoskirsch): Don't care, except for documentation
+                'documentation': {
+                'description': 'Array of data for the new arbitrary waveform. This may be an iterable of float, or for best performance a numpy.ndarray of dtype int16 or float64.',
+                },
+            },
+            {
+                'direction': 'out',
+                'enum': None,
+                'name': 'waveformHandle',
+                'type': 'ViInt32',
+                'documentation': {
+                    'description': 'The handle that identifies the new waveform. This handle is used in other methods when referring to this waveform.',
+                },
+            },
+        ],
+'documentation': {
+'description': '''
+Creates an onboard waveform
+for use in Arbitrary Waveform output mode or Arbitrary Sequence output
+mode.
+''',
+'note': '''
+You must set NIFGEN\_ATTR\_OUTPUT\_MODE to NIFGEN\_VAL\_OUTPUT\_ARB or
+NIFGEN\_VAL\_OUTPUT\_SEQ before calling this function.
+''',
+},
+    },
+
     'WriteWaveformDispatcher': {
         'codegen_method': 'public',
         'returns': 'ViStatus',
@@ -152,8 +200,8 @@ functions_additional_functions = {
                 'name': 'waveformNameOrHandle',
                 'type': 'ViInt32',  #TODO(marcoskirsch): Don't care, except for documentation
                 'documentation': {
-                'description': 'The name (str) or handle (int) of an arbitrary waveform previously allocated with the niFgen\_AllocateNamedWaveform niFgen\_AllocateWaveform function.',
-            },
+                    'description': 'The name (str) or handle (int) of an arbitrary waveform previously allocated with niFgen\_AllocateNamedWaveform or niFgen\_AllocateWaveform.',
+                },
             },
             {
                 'direction': 'in',
@@ -165,7 +213,7 @@ functions_additional_functions = {
                 'direction': 'in',
                 'enum': None,
                 'name': 'Data',
-                'type': 'ViReal64[]',
+                'type': 'ViReal64[]',  #TODO(marcoskirsch): Don't care, except for documentation
                 'documentation': {
                     'description': 'Array of data to load into the waveform. This may be an iterable of float, or for best performance a numpy.ndarray of dtype int16 or float64.',
                 },
@@ -185,11 +233,21 @@ nifgen\_SetWaveformNextWritePosition function.''',
 # Override the 'python' name for some functions.
 functions_python_name = {
     'AbortGeneration':                      { 'python_name': 'abort',                   },
-    'CreateWaveformF64':                    { 'python_name': 'create_waveform',         },
+    'CreateWaveformDispatcher':             { 'python_name': 'create_waveform'          },
     'WriteWaveformDispatcher':              { 'python_name': 'write_waveform'           },
 }
 
 functions_method_template_filenames = {
+    'CreateWaveformDispatcher':     { 'method_templates': [
+        { 'session_filename': 'create_waveform', 'documentation_filename': 'default_method', 'method_python_name_suffix': '_numpy', },
+    ], },
+    'CreateWaveformF64':            { 'method_templates': [
+        { 'session_filename': 'default_method', 'documentation_filename': 'default_method', 'method_python_name_suffix': '', },
+        { 'session_filename': 'numpy_write_method', 'documentation_filename': 'numpy_method', 'method_python_name_suffix': '_numpy', },
+    ], },
+    'CreateWaveformI16':            { 'method_templates': [
+        { 'session_filename': 'numpy_write_method', 'documentation_filename': 'numpy_method', 'method_python_name_suffix': '_numpy', },
+    ], },
     'WriteWaveformDispatcher':      { 'method_templates': [
         { 'session_filename': 'write_waveform', 'documentation_filename': 'default_method', 'method_python_name_suffix': '', },
     ], },
@@ -198,7 +256,7 @@ functions_method_template_filenames = {
         { 'session_filename': 'numpy_write_method', 'documentation_filename': 'numpy_method', 'method_python_name_suffix': '_numpy', },
     ], },
     'WriteNamedWaveformF64':        { 'method_templates': [
-        { 'session_filename': 'default_method', 'documentation_filename': 'default_method', 'method_python_name_suffix': '', }, 
+        { 'session_filename': 'default_method', 'documentation_filename': 'default_method', 'method_python_name_suffix': '', },
         { 'session_filename': 'numpy_write_method', 'documentation_filename': 'numpy_method', 'method_python_name_suffix': '_numpy', },
     ], },
     'WriteBinary16Waveform':        { 'method_templates': [
@@ -210,9 +268,10 @@ functions_method_template_filenames = {
 }
 
 functions_numpy = {
+    'CreateWaveformF64':            { 'parameters': { 3: { 'numpy': True, }, }, },
+    'CreateWaveformI16':            { 'parameters': { 3: { 'numpy': True, }, }, },
     'WriteWaveform':                { 'parameters': { 4: { 'numpy': True, }, }, },
     'WriteNamedWaveform':           { 'parameters': { 4: { 'numpy': True, }, }, },
     'WriteBinary16Waveform':        { 'parameters': { 4: { 'numpy': True, }, }, },
     'WriteNamedWaveformI16':        { 'parameters': { 4: { 'numpy': True, }, }, },
 }
-